chore: move functions to src folder
